feat: Set up router and render components.
Import the 'createBrowserRouter' and 'RouterProvider' from the 'react-router-dom' library in the index.js file.
Set up the router using the 'createBrowserRouter' function, specifying paths, elements, children paths and
Wrap the root application with the 'RouterProvider' component passing the created router as a props in the index.js file.
Reder the 'Header' and 'Outle' that render component base on route defined in router in the App.js.