feat: Reset or remain input value.
Implement a logic to either reset or remain the 'searchKeyword' state base on the current path.
Use the 'useEffect' hook to listen changes in the current path.
Remain the 'searchKeyword' state by not updating it when the current path matches '/videos/:keyword'.
Reset the 'searchKeyword' state by setting it to an empty string when current path does not match '/videos/:keyword'.