feat: Add the 'related' and 'detail' types.
Add the 'related' type to fetch the 'related' videos base on the 'videoId' from the router parameter.
Add the 'detail' tyep to fetch the 'detail' vodeo base on the 'videoId' from the router parameter.
Add the 'video' state using the 'useState' hook.
If the value of type is 'detail', return the 'video' state.
Otherwise, return the 'videos' state.