feat: Render Detail component.
Fetch the 'relatedVideos' and 'video' data using the 'useFetchVideos' hook.
Create a new 'VideoPlaybox' comopnent to render the iframe and infromation of the clicked video.
Render the 'VideoPlayBox' component by passing the 'video' as props in the 'Detail' component.
Render the 'Video' component by passing the 'relatedVideos' as props in the 'Detail' component.