feat: Create fake api class.
Create a new file called the 'fake-youtube-api.js' and define the 'FakeYoutubeApi' class.
Add a seach function to the 'FakeYoutubeApi' class that returns a private function base on the keyword.
Add private functions to the 'FakeYoutubeApi' class that return fake data.
Impelement the 'FakeYoutube' class and its functions in the 'queryFn' argument within the 'useQuery' hook in the 'Home' component.